# doc/api.yaml
openapi: 3.0.0
info:
  title: Fantastic Coffee (decaffeinated)
  version: 1.0.0
servers:
  - url: 'https://localhost:3000'
...
components:
  schemas:
    UserID:
      description: The unique id of a user, automatically generated when the account is created.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
...
    MessageID:
      description: The unique identifier of a message, automatically generated when the message is first created and sent.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
...
    ConversationID:
      description: The unique identifier of a chat, automatically generated when the conversation is started.
      type: integer
      format: int32
      minimum: 1
      maximum: 100000
      readOnly: true
...
    Emoji:
      description: an emoji
      type: string
      pattern: "[\u203C-\u3299\u1F004\u1F0CF\u1F9FF]"
      minLength: 0
      maxLength: 1
      default: ""
    Reaction: 
      description: The reaction/emoji to a messsage; each person can put at most one emoji.
      type: object
      properties:
        sender:
          $ref: "#/components/schemas/UserID"
        emoji:
          $ref: "#/components/schemas/Emoji"
...
  parameters:
...
    UserId:
      description: The user identifier, which must be passed for any possible operation.
      name: usrId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/UserID"
    ConversationId:
      description: The identifer of a conversation, used for any kind of acess to that conversation.
      name: convId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ConversationID"
    MessageId:
      description: The identifer of a message, used whenever you want to comment, respond, share, or modify that message.
      name: msgId
      in: path
      required: true
      schema: 
        $ref: "#/components/schemas/MessageID"
  securitySchemes:

    AuthenticationId:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid.
    ServerError: 
      description: Server error occurred. Please, try again in a few minutes.
    NotFoundError:
      description: The given id(s) doesn't correspond witha any user/session/message. Please try again.
    BadRequestError:
      description: You either aren't an admin to do this operation or the given input(s) are invalid.
    ItemDeleted:
      description: Item(s) succesfully deleted.
...
security:
  - AuthenticationId: []
paths:
...
  /users/{usrId}/conversations/{convId}/messages/{msgId}/reactions:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/ConversationId"
      - $ref: "#/components/parameters/MessageId"
    post: 
      tags: [message]
      summary: comments a message
      description: |-
        given a message in the chat, the user can comment it
        (a.k.a react to it) by adding the desired emoji to the message's
        reactions pool.
        If an emoji associated to your Id is already present in
        the message's reaction pool, the one in the pool will be replaced with the newly chosen one.
        If everything goes well, the updated message (with the updated reaction pool) is returned.
      operationId: commentMessage
      requestBody:
        content:
          text/plain:
            schema:
              $ref: "#/components/schemas/Emoji"
      responses:
        "201":
          description: Reaction sent or changed succesfully. 
          content:
            application/json:
              schema: 
                $ref: "#/components/schemas/Reaction"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [message]
      summary: removes a reaction.
      description: |-
        checks for a reaction having the same sender id as the user,
        in which case, the reaction sent by the user for that specific message will be removed.
      operationId: uncommentMessage
      responses:
        "204": 
          $ref: "#/components/responses/ItemDeleted"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/ServerError"